requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
environs==11.0.0
marshmallow==3.21.2
packaging==24.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
